[shader_recompiler/Maxwell] Fix typo in ISBERD Bitfield

+++ src/shader_recompiler/frontend/maxwell/translate/impl/internal_stage_buffer_entry_read.cpp
@@ -28,8 +28,8 @@ void TranslatorVisitor::ISBERD(u64 insn) {
         u64 raw;
         BitField<0, 8, IR::Reg> dest_reg;
         BitField<8, 8, IR::Reg> src_reg;
-        BitField<8, 8, u32> src_reg_num;
-        BitField<24, 8, u32> imm;
+        BitField<8, 8, u64> src_reg_num;
+        BitField<24, 8, u64> imm;
         BitField<31, 1, u64> skew;
         BitField<32, 1, u64> o;
         BitField<33, 2, Mode> mode;
